構造見直し完了：不要ファイル削除

+++ src/main/java/xyz/skycat/work/VelocityTemplateGen/ae/util/FilesUtil.java
@@ -13,7 +13,9 @@ public class FilesUtil {
 
     public static boolean outputInit(String vmFileNameFullPath) {
         try {
-            Files.deleteIfExists(Paths.get(vmFileNameFullPath));
+            if (Files.exists(Paths.get(vmFileNameFullPath))) {
+                Files.deleteIfExists(Paths.get(vmFileNameFullPath));
+            }
             Files.createFile(Paths.get(vmFileNameFullPath));
         } catch (IOException e) {
             // TODO error handling.
